# server/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Alure API
  version: 0.1.0
servers:
  - url: /api/v1
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Login and obtain a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /auth/bootstrap:
    get:
      tags: [auth]
      summary: Check if an administrator account exists
      responses:
        '200':
          description: Bootstrap status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapStatus'
    post:
      tags: [auth]
      summary: Create the initial administrator account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapRequest'
      responses:
        '200':
          description: Bootstrap success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /auth/me:
    get:
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
    put:
      tags: [auth]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfile'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  /auth/users:
    get:
      tags: [auth]
      summary: List users (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserListItem'
    post:
      tags: [auth]
      summary: Create a user invite (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: Invite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
  /auth/users/{userId}/resend:
    post:
      tags: [auth]
      summary: Resend a user invite (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invite resent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResendInviteResponse'
  /auth/invite/accept:
    post:
      tags: [auth]
      summary: Accept an invite and set a password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInviteRequest'
      responses:
        '200':
          description: Invite accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptInviteResponse'
  /auth/invite/status:
    get:
      tags: [auth]
      summary: Check invite status
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invite status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteStatusResponse'
  /licenses/activate:
    post:
      tags: [licensing]
      summary: Activate a license key and issue a signed receipt
      security:
        - bearerAuth: [licensing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateRequest'
      responses:
        '200':
          description: Activation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivateResponse'
  /licenses/verify:
    post:
      tags: [licensing]
      summary: Verify a receipt and optionally refresh it
      security:
        - bearerAuth: [licensing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
  /licenses/revoke:
    post:
      tags: [licensing]
      summary: Revoke a license or a specific activation
      security:
        - bearerAuth: [licensing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeRequest'
      responses:
        '200':
          description: Revoke result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeResponse'
  /licenses:
    get:
      tags: [licensing]
      summary: List licenses (optionally filtered by project)
      parameters:
        - in: query
          name: project_id
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Licenses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LicenseListItem'
    post:
      tags: [licensing]
      summary: Create a new license key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLicenseRequest'
      responses:
        '200':
          description: License created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateLicenseResponse'
  /licenses/bulk:
    post:
      tags: [licensing]
      summary: Create multiple licenses and email recipients
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkCreateLicensesRequest'
      responses:
        '200':
          description: Bulk license result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkCreateLicensesResponse'
  /licenses/{license_id}/activations:
    get:
      tags: [licensing]
      summary: List activations for a license
      parameters:
        - in: path
          name: license_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Activation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivationListItem'
  /licenses/{license_id}/modules:
    get:
      tags: [licensing]
      summary: List modules for a license
      parameters:
        - in: path
          name: license_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: License modules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseModulesResponse'
    post:
      tags: [licensing]
      summary: Update modules enabled for a license
      parameters:
        - in: path
          name: license_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLicenseModulesRequest'
      responses:
        '200':
          description: Updated license modules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseModulesResponse'
  /licenses/{license_id}/activations/{activation_id}/modules:
    get:
      tags: [licensing]
      summary: List modules for an activation
      parameters:
        - in: path
          name: license_id
          required: true
          schema:
            type: string
        - in: path
          name: activation_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Activation modules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivationModulesResponse'
    post:
      tags: [licensing]
      summary: Update modules enabled for an activation
      parameters:
        - in: path
          name: license_id
          required: true
          schema:
            type: string
        - in: path
          name: activation_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActivationModulesRequest'
      responses:
        '200':
          description: Updated activation modules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivationModulesResponse'
  /licenses/{license_id}/activations/{activation_id}/hostname:
    post:
      tags: [licensing]
      summary: Reveal activation hostname (requires password)
      parameters:
        - in: path
          name: license_id
          required: true
          schema:
            type: string
        - in: path
          name: activation_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevealActivationHostnameRequest'
      responses:
        '200':
          description: Hostname
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevealActivationHostnameResponse'
...
  /plans:
    get:
      tags: [plans]
      summary: List plans
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Plans list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plan'
    post:
      tags: [plans]
      summary: Create a plan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlan'
      responses:
        '200':
          description: Plan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
  /plans/{plan_id}:
    put:
      tags: [plans]
      summary: Update a plan
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: plan_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlan'
      responses:
        '200':
          description: Plan updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
    delete:
      tags: [plans]
      summary: Delete a plan
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: plan_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ActivateRequest:
      type: object
      required: [license_key, device_id]
      properties:
        license_key:
          type: string
        device_id:
          type: string
        app_version:
          type: string
        device_meta:
          type: object
          additionalProperties: true
    ActivateResponse:
      type: object
      required: [receipt, activation_id, grace_period_days, server_time]
      properties:
        receipt:
          type: string
          description: Signed receipt token
        activation_id:
          type: string
        expires_at:
          type: string
          format: date-time
        grace_period_days:
          type: integer
        server_time:
          type: string
          format: date-time
    VerifyRequest:
      type: object
      required: [receipt, device_id]
      properties:
        receipt:
          type: string
        device_id:
          type: string
    VerifyResponse:
      type: object
      required: [valid, server_time]
      properties:
        valid:
          type: boolean
        reason:
          type: string
        new_receipt:
          type: string
        expires_at:
          type: string
          format: date-time
        server_time:
          type: string
          format: date-time
    RevokeRequest:
      type: object
      properties:
        license_id:
          type: string
        activation_id:
          type: string
        reason:
          type: string
      description: Provide license_id or activation_id
    RevokeResponse:
      type: object
      required: [revoked]
      properties:
        revoked:
          type: boolean
        server_time:
          type: string
          format: date-time
    CreateLicenseRequest:
      type: object
      required: [project_id, plan, max_activations]
      properties:
        project_id:
          type: string
        plan:
          type: string
        max_activations:
          type: integer
          minimum: 0
        duration_days:
          type: integer
        notes:
          type: string
        module_keys:
          type: array
          items:
            type: string
    CreateLicenseResponse:
      type: object
      required: [license_id, license_key]
      properties:
        license_id:
          type: string
        license_key:
          type: string
    BulkCreateLicensesRequest:
      type: object
      required: [project_id, plan, max_activations, recipients]
      properties:
        project_id:
          type: string
        plan:
          type: string
        max_activations:
          type: integer
          minimum: 0
        duration_days:
          type: integer
        notes:
          type: string
        recipients:
          type: array
          items:
            type: string
        module_keys:
          type: array
          items:
            type: string
    BulkCreateLicenseItem:
      type: object
      required: [email, license_id, license_key]
      properties:
        email:
          type: string
        license_id:
          type: string
        license_key:
          type: string
    BulkCreateLicenseError:
      type: object
      required: [email, error]
      properties:
        email:
          type: string
        error:
          type: string
        license_id:
          type: string
        license_key:
          type: string
    BulkCreateLicensesResponse:
      type: object
      required: [created, failed]
      properties:
        created:
          type: array
          items:
            $ref: '#/components/schemas/BulkCreateLicenseItem'
        failed:
          type: array
          items:
            $ref: '#/components/schemas/BulkCreateLicenseError'
    LicenseListItem:
      type: object
      required: [license_id, project_id, plan, max_activations, revoked, created_at]
      properties:
        license_id:
          type: string
        project_id:
          type: string
        plan:
          type: string
        max_activations:
          type: integer
          minimum: 0
        revoked:
          type: boolean
        duration_days:
          type: integer
        notes:
          type: string
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    ActivationListItem:
      type: object
      required: [activation_id, device_id_hash, revoked, created_at]
      properties:
        activation_id:
          type: string
        device_id_hash:
          type: string
        revoked:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        hostname_masked:
          type: string
    ActivationModuleItem:
      type: object
      required: [module_id, key, name, enabled]
      properties:
        module_id:
          type: string
        key:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        force_activation:
          type: boolean
        force_deactivation:
          type: boolean
        params:
          type: object
          additionalProperties: true
    ActivationModulesResponse:
      type: object
      required: [modules]
      properties:
        modules:
          type: array
          items:
            $ref: '#/components/schemas/ActivationModuleItem'
    UpdateActivationModulesRequest:
      type: object
      required: [module_keys]
      properties:
        module_keys:
          type: array
          items:
            type: string
    RevealActivationHostnameRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
    RevealActivationHostnameResponse:
      type: object
      properties:
        hostname:
          type: string
    LicenseModuleItem:
      type: object
      required: [module_id, key, name, enabled]
      properties:
        module_id:
          type: string
        key:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        force_activation:
          type: boolean
        force_deactivation:
          type: boolean
        params:
          type: object
          additionalProperties: true
    LicenseModulesResponse:
      type: object
      required: [modules]
      properties:
        modules:
          type: array
          items:
            $ref: '#/components/schemas/LicenseModuleItem'
    UpdateLicenseModulesRequest:
      type: object
      required: [modules]
      properties:
        modules:
          type: array
          items:
            type: object
            required: [key]
            properties:
              key:
                type: string
              force_activation:
                type: boolean
              force_deactivation:
                type: boolean
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
    UpdateProfile:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    CreateUserRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
    CreateUserResponse:
      type: object
      required: [user_id, email, invite_expires_at]
      properties:
        user_id:
          type: string
        email:
          type: string
        invite_expires_at:
          type: string
          format: date-time
    AcceptInviteRequest:
      type: object
      required: [token, otp, password]
      properties:
        token:
          type: string
        otp:
          type: string
        password:
          type: string
    AcceptInviteResponse:
      type: object
      required: [accepted]
      properties:
        accepted:
          type: boolean
    InviteStatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [valid, expired, used, invalid]
    UserListItem:
      type: object
      required: [user_id, email, name, role, invite_status]
      properties:
        user_id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
        invite_status:
          type: string
          enum: [accepted, pending, expired]
        invite_expires_at:
          type: string
          format: date-time
    ResendInviteResponse:
      type: object
      required: [user_id, invite_expires_at]
      properties:
        user_id:
          type: string
        invite_expires_at:
          type: string
          format: date-time
    BootstrapStatus:
      type: object
      required: [has_admin]
      properties:
        has_admin:
          type: boolean
    BootstrapRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
    Plan:
      type: object
      required: [id, name, allowed_channels, max_activations_default, grace_period_days, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        allowed_channels:
          type: array
          items:
            type: string
        max_activations_default:
          type: integer
          minimum: 0
        grace_period_days:
          type: integer
        duration_days_default:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreatePlan:
      type: object
      required: [name, allowed_channels, max_activations_default, grace_period_days]
      properties:
        name:
          type: string
        allowed_channels:
          type: array
          items:
            type: string
        max_activations_default:
          type: integer
          minimum: 0
        grace_period_days:
          type: integer
        duration_days_default:
          type: integer
    UpdatePlan:
      type: object
      properties:
        name:
          type: string
        allowed_channels:
          type: array
          items:
            type: string
        max_activations_default:
          type: integer
          minimum: 0
        grace_period_days:
          type: integer
        duration_days_default:
          type: integer
...